refactor libk
