+++ kernel/include/canvas.h
