+++ kernel/include/color.h
