+++ kernel/include/terminal.h
