+++ kernel/src/c/graphics/canvas.c
