+++ kernel/src/c/terminal/initialize.c
