+++ kernel/src/c/terminal/terminal_state.c
