+++ kernel/src/c/terminal/write.c
